+++ dataduke.txt
@@ -1,2 +1,1 @@
-todo return book
-deadline return book by/ June 6th
+
