+++ src/main/java/Event.java
@@ -1,15 +1,24 @@
+import java.time.*;
+import java.time.format.DateTimeFormatter;
+
 public class Event extends Task{
 
-    protected String at;
+    protected LocalDate date;
 
-    public Event(String description, String at) {
-        super(description);
-        this.at = at;
+    public Event(String description, LocalDate at, String enter) {
+        super(description, enter);
+        this.date = at;
     }
 
     @Override
-    public String toString() {
-        return " [E]" + "["+getStatusIcon()+"] "+getDescription() + " (at: " + at + ")";
+    public LocalDate getDate() {
+        return this.date;
+    }
+
+    @Override
+
+    public String printTask() {
+        return "[E]" + "["+getStatusIcon()+"] "+getDescription() + " (at: " + date.format(DateTimeFormatter.ofPattern("MMM/dd/yyyy")) + ")";
     }
 
 }
